SOMA for ANN training

+++ NAVY/ANN/Connection.cs
@@ -7,9 +7,9 @@ namespace ANN
     {
         public Neuron From { get; private set; }
         public Neuron To { get; private set; }
-        public double Weight { get; set; }
+        public float Weight { get; set; }
 
-        public Connection (Neuron from, Neuron to, double weight)
+        public Connection (Neuron from, Neuron to, float weight)
         {
             From = from;
             To = to;
